python_wrapper/example/requirements.txt
numpy~=1.21.0
matplotlib~=3.4.2
scipy~=1.6.3
umap-learn~=0.5.1
nptsne~=1.2.0
scikit-learn~=0.24.2
imageio~=2.13.0